requirements.txt
aggdraw==1.3.12
backcall==0.2.0
certifi==2021.5.30
cftime==1.5.0
chardet==4.0.0
configobj==5.0.6
cycler==0.10.0
decorator==5.0.9
idna==2.10
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
netCDF4==1.5.6
numpy==1.20.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
PIL-Tools==1.1.0
Pillow==8.2.0
pkg-resources==0.0.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycoast==1.4.0
Pygments==2.9.0
pykdtree==1.3.4
pyparsing==2.4.7
pyproj==3.1.0
pyresample==1.20.0
pyshp==2.1.3
python-dateutil==2.8.1
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
six==1.16.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.2.5
